# Conflicts:
#	src/main/java/com/utilityhub/apartmentutilityhub/controller/ApartmentController.java
#	src/main/java/com/utilityhub/apartmentutilityhub/service/ApartmentService.java
#	src/main/java/com/utilityhub/apartmentutilityhub/service/impl/ApartmentServiceImpl.java
#	src/main/resources/templates/apartment-list.html
#	src/main/resources/templates/layout.html

+++ src/main/java/com/utilityhub/apartmentutilityhub/service/impl/ApartmentServiceImpl.java
@@ -49,7 +49,7 @@ public class ApartmentServiceImpl {
     }
 
     // Delete apartment by ID
-    public void deleteApartmentById(Integer id) {
+    public void deleteApartmentById(Long id) {
         apartmentRepo.deleteById(id);
     }
